backend/requirements.txt
asgiref==3.7.2
click==8.1.7
Django==3.2.23
django_debug_toolbar==3.8.1
Flask==2.2.5
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
